feat: add pinia to use the store in the project
