+++ package.json
@@ -10,7 +10,9 @@
     "postinstall": "nuxt prepare"
   },
   "dependencies": {
+    "@pinia/nuxt": "^0.9.0",
     "nuxt": "^3.14.1592",
+    "pinia": "^2.3.0",
     "vue": "latest",
     "vue-router": "latest"
   },
