+++ src/components/general/DFooterNavBar.vue
@@ -1,36 +1,7 @@
 <script setup lang="ts">
-const navItems = ref([
-    {
-        active: false,
-        id: '1',
-        title: 'Home',
-        route: '/',
-    },
-    {
-        active: false,
-        id: '2',
-        title: 'Services',
-        route: '/services',
-    },
-    {
-        active: false,
-        id: '3',
-        title: 'Blogs',
-        route: '/blogs',
-    },
-    {
-        active: false,
-        id: '4',
-        title: 'About',
-        route: '/about',
-    },
-    {
-        active: false,
-        id: '5',
-        title: 'Contacts',
-        route: '/contacts',
-    }
-])
+import { useNavStore } from '~/store/useNavStore'
+
+const { navItems } = storeToRefs(useNavStore())
 </script>
 
 <template>
